requirements.txt
hyperopt
sklearn
tradingview_ta
pydantic
pandas
xgboost
dotenv
python-telegram-bot
matplotlib
tabulate